Add example for code-splitting
