+++ examples/django-react/package.json
@@ -4,7 +4,8 @@
   "private": true,
   "dependencies": {
     "react": "^15.6.1",
-    "react-dom": "^15.6.1"
+    "react-dom": "^15.6.1",
+    "react-loadable": "^4.0.5"
   },
   "scripts": {
     "start": "BROWSER=false react-scripts start",
