+++ examples/django-react/src/core/App.css
