+++ examples/django-react/src/core/App.test.js
