+++ examples/django-react/src/core/CodeSplit.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+class CodeSplit extends Component {
+  render() {
+    return (
+      <div>
+        <h1>Code Split Example</h1>
+        <p>This component is loaded dynamically</p>
+      </div>
+    );
+  }
+}
+
+export default CodeSplit;
